Ensure period is set within the hardware's limitations.
It is possible for an app to send out a notification that has a period of less than 2 seconds. When that happens, the period is not set at all and causes strange things to happpen, including the trackball not flashing at all.

Thsi commit simply ensures that period is always between 2 and 12.